// src/shared/api/generated/model/getDiscoveryCreatorsGender.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */

export type GetDiscoveryCreatorsGender = typeof GetDiscoveryCreatorsGender[keyof typeof GetDiscoveryCreatorsGender];


export const GetDiscoveryCreatorsGender = {
  male: 'male',
  female: 'female',
  non_binary: 'non_binary',
} as const;